[modify]分析カードを1つのコンテナに統一
